Discard reserved register bits on reads
When reading slave status and data, clear the bits of reserved positions.
Also, change integer sizes - status is 8-bit, data is 32.

+++ src/i2cinterruptcontext.hpp
@@ -4,7 +4,7 @@
 #include <cstdint>
 
 struct I2CInterruptContext {
-    const std::uint_fast32_t slaveStatus;
+    const std::uint_fast8_t slaveStatus;
     const std::uint_fast32_t slaveData;
 };
